requirements.txt
alive_progress==3.1.5
arrow==1.3.0
ddddocr==1.5.3
loguru==0.7.2
lxml==5.3.0
Pillow==10.4.0
playwright==1.44.0
prettytable==3.11.0
Requests==2.32.3
